# specs/002-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Web Application API
  description: REST API for the Full-Stack Todo Web Application with JWT authentication and user-isolated data
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

tags:
  - name: auth
    description: Authentication operations (no JWT required)
  - name: tasks
    description: Task management operations (JWT required)

paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      description: Creates a new user account with email and password. Returns JWT token upon successful registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user
      description: Authenticates user with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout user
      description: Invalidates the user's session (stateless - no server-side session management)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks:
    get:
      tags:
        - tasks
      summary: Get user's tasks
      description: Retrieve all tasks for the authenticated user. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - tasks
      summary: Create a new task
      description: Creates a new task for the authenticated user. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks/{id}:
    get:
      tags:
        - tasks
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The task ID
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGetResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - tasks
      summary: Update a task
      description: Update an existing task for the authenticated user. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskUpdateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - tasks
      summary: Delete a task
      description: Delete a specific task for the authenticated user. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks/{id}/complete:
    patch:
      tags:
        - tasks
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task. User ID in JWT must match the user_id in the path.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user ID (must match the authenticated user's ID)
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskToggleRequest'
      responses:
        '200':
          description: Task status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskUpdateResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user ID in token does not match user_id in path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegistrationRequest:
      type: object
      required:
        - email
        - password
        - password_confirm
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        full_name:
          type: string
          maxLength: 100
          example: "John Doe"
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: "SecurePassword123!"
        password_confirm:
          type: string
          minLength: 8
          maxLength: 128
          example: "SecurePassword123!"

    UserRegistrationResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/UserToken'
        error:
          type: string
          nullable: true

    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "SecurePassword123!"

    UserLoginResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/UserToken'
        error:
          type: string
          nullable: true

    UserToken:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          default: "bearer"
        user:
          $ref: '#/components/schemas/UserPublic'

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          example: "Milk, bread, eggs, cheese"

    TaskCreateResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/TaskPublic'
        error:
          type: string
          nullable: true

    TaskGetResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/TaskPublic'
        error:
          type: string
          nullable: true

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
          example: "Updated task title"
        description:
          type: string
          nullable: true
          example: "Updated description"
        is_completed:
          type: boolean
          nullable: true
          example: true

    TaskUpdateResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/TaskPublic'
        error:
          type: string
          nullable: true

    TaskToggleRequest:
      type: object
      required:
        - is_completed
      properties:
        is_completed:
          type: boolean
          example: true

    TaskListResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          default: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskPublic'
        error:
          type: string
          nullable: true

    UserPublic:
      type: object
      required:
        - id
        - email
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        full_name:
          type: string
          nullable: true
          example: "John Doe"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    TaskPublic:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          example: "Milk, bread, eggs, cheese"
        is_completed:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          default: false
        data:
          type: object
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Task title is required"
            details:
              type: object
              nullable: true